Shrink the size of the Meta struct. (#48)
* Shrink the size of the Meta struct.

The size of the Any enum kept growing as Meta gets added to more places.
The alternative would be to Box these optional fields, but I think a Vec
better captures the arbitrary box nature. A map would be nice too but
it's not possible to type K:V.

before main: 816 bytes
after main: 528 bytes

before #47: 1168 bytes
after  #47: 624 bytes

The only thing I don't like about this approach is Any. There's a lot of
boxes that we know *should* never be within a meta box. However, it's a lot
more work to create a MetaAny enum and frankly, not worth it.

* clippy