fix!: trun within traf should be Vec not Option
Section 8.8.8.1 defines the Track fragment run box as "Quantity: Zero or
more", rather than "Zero or one".

We can see an example of multiple trun boxes in a traf with this example
segment:
https://github.com/MPEGGroup/FileFormatConformance/blob/main/data/file_features/published/isobmff/10_fragments.mp4

BREAKING CHANGE: The type of the `Traf::trun` property has changed from
`Option<Trun>` to `Vec<Trun>`.