fix: limit capacity of vecs against attack vector
PR review had a very valid point that using `Vec::with_capacity` with a
number that is provided by an untrusted source is dangerous, as it is
easy for a malicious actor to define the number to be very high, and
cause the program to allocate massive amounts of memory.

I chose to still use `with_capacity`, but with a limiter of 1024, as
I've seen a few examples of segments with hundreds of entries for each
of the 3 new boxes. I believe that currently Rust will double the
allocated memory on each resize, so, for example, I believe that means
that for a segment with 399 entries (example I found), could cause about
9 resizes, which seems nice to try and avoid.

Also used `as usize` rather than `usize::try_from` as this should be
safe across all CPU targets where this library could be used.