fix: encode AvccExt masked properties correctly
While validating encode of Avcc I found that the reserved bits were not
being re-added to the ext properties in Avcc. This change adds them back
in for encode and also subtracts 8 from the properties where 8 was added
in.

The "throw error if out of range" approach follows the precedent set by
the `length_size` encoding above, but note, this is a change in logic.
This may also seem contradictory to add into a PR where we are removing
strictness in the decode; however, I think that this is different,
because this couldn't fail when the property is derived from decoded
data because the reserved bits are lost, and so this is only catching
bad input when a user is constructing the atom directly.