Relaxed reserved bits validation on `Avcc` decode (#55)
* feat: loosen reserved bits validation in avcc

Some example streams, that work fine on many clients, were failing to
decode, because the reserved bits were each set to 0 instead of 1. While
this is incorrect, becuase it works with many player implementations, it
is seen as OK to reduce the strictness of the decode validation.

Fixes #54

* fix: encode AvccExt masked properties correctly

While validating encode of Avcc I found that the reserved bits were not
being re-added to the ext properties in Avcc. This change adds them back
in for encode and also subtracts 8 from the properties where 8 was added
in.

The "throw error if out of range" approach follows the precedent set by
the `length_size` encoding above, but note, this is a change in logic.
This may also seem contradictory to add into a PR where we are removing
strictness in the decode; however, I think that this is different,
because this couldn't fail when the property is derived from decoded
data because the reserved bits are lost, and so this is only catching
bad input when a user is constructing the atom directly.

* fix: better test input value

Co-authored-by: Brad Hards <bradh@frogmouth.net>

* fix: better expect failure message

Co-authored-by: Brad Hards <bradh@frogmouth.net>

---------

Co-authored-by: Brad Hards <bradh@frogmouth.net>