implement basic FLAC support
This is based on RFC9639 and its outreference to https://github.com/xiph/flac/blob/master/doc/isoflac.txt

The metadata generation and parsing is incomplete. It does do the Streaminfo which is the critical part.

Also, the u24 implementation is still a bit ugly. Would appreciate recommendations for that.