fix(eac3): uses the correct number of sub streams, consumes all remaining bytes by definition (#61)
* Adds more comprehensive eac3 tests

* fix(eac3): fixes off-by-one error in EAC3 substream count

Corrects an error in the calculation of the number of substreams in the EAC3 specific box.

The number of substreams was incorrectly incremented during decoding and not decremented during encoding.

Updates test sample to be correctly counted as well.

* fix(eac3): consumes any remaining reserved data

Ensures that any reserved bits following the valid data within the `Ec3SpecificBox` are properly consumed, preventing potential parsing errors, i.e. `Error::UnderDecode`